requirements.txt
Django==2.1.2
django-extensions==2.1.3
djangorestframework==3.8.2
pkg-resources==0.0.0
python-decouple==3.1
pytz==2018.5
six==1.11.0
